Car can now move and turn correctly (and drawn correctly) Need to redo spritesheets to make 0 degrees facing Right and 90 North, current code comensates for it but this needs to be changed for consistency. TODO: add physics to the car moving

+++ BumperCars/Car.h
@@ -34,16 +34,16 @@ private:
 
     const int PI = 3.14159265359;
     ///positional
-    int posX;
-    int posY;
+    double posX;
+    double posY;
     int angle;
-    int velocity;
-    int acceleration;
+    double velocity;
+    double acceleration;
 
     ///properties
-    int maxSpeed;
-    int maxBackThrust;
-    int rotateSpeed;
+    double maxSpeed;
+    double maxBackThrust;
+    double rotateSpeed;
     int maxHealth;
     int mass;
